rm register_*_domain and mv func to *_domain.go
